feat: Add dice screen and cubit
This commit adds the `DiceScreen` widget and `DiceCubit` class to the codebase. The `DiceScreen` widget displays an image of a dice based on the current state of the `DiceCubit`. The `DiceCubit` is responsible for generating random dice values when the `roll()` method is called.

The changes also include updating the `pubspec.yaml` file to add the `flutter_bloc` and `bloc_test` dependencies.

This commit addresses the need to display a dice image and generate random dice values for the dice rolling app.